requirements.txt
altair
numpy
pandas
pydeck
streamlit
google-generativeai
pdf2image